updating

+++ 12 - array/fibonacci.c
@@ -7,7 +7,7 @@ int main() {
     printf("n: ");
     scanf("%d", &n);
 
-    int nums[n];
+    long long nums[n];
 
     nums[0] = 0;
     nums[1] = 1;
@@ -16,7 +16,7 @@ int main() {
         nums[i] = nums[i-2] + nums[i-1];
     }
     for (int i=0; i<n; i++) {
-        printf("Num %d = %d\n", i+1, nums[i]);
+        printf("Num %d = %lld\n", i+1, nums[i]);
     }
 
     return 0;
